+++ CHANGELOG.md
@@ -77,6 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   which have formerly not been invoked by the controller. (#207)
 - `org.somda.sdc.glue.consumer.report.helper.ReportWriter` no longer fails on descriptors without states in description modification report parts with modification type del. (#210)
 - Disabled external XML entity processing to prevent XXE attacks. (#218)
+- `org.somda.sdc.dpws.soap.wsdiscovery.WsDiscoveryUtil` correctly compares scope sets using STRCMP0 rules. (#217)
 
 ## [1.1.0] - 2020-04-18